requirements.txt
pandas==2.1.3
numpy==1.26.2 
openai
sklearn
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
db-dtypes>=1.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv>=1.0.0
requests>=2.31.0
